+++ src/ReleaseNotes.c
@@ -20,7 +20,23 @@ struct ReleaseInfo
 //    {   "Title",
 //            "Descrition"},
 //};
-static struct ReleaseNote m_Notes_NEXT[]=
+static struct ReleaseNote m_Notes_Rel1_0_2_0[]=
+{
+    {   "Fixed problem with Linux installer",
+            "The Linux installer was missing a depends for the QT multimedia.  Added it and added version numbers to the packages."},
+    {   "Fixed bug with open new connection from URI dialog",
+            "Dialog was not setting up options which depending on the device would prevent connection from working."},
+    {   "Fixed bug with importing settings does not apply the settings",
+            "When you imported new settings the settings where not being "
+            "applied to connections, but new connections would get the "
+            "settings."},
+};
+static struct ReleaseNote m_Notes_Rel1_0_1_0[]=
+{
+    {   "Fixed bug with plugins with long descriptions",
+            "The wrong buffer was used when loading a plugin with a description longer than 99 bytes"},
+};
+static struct ReleaseNote m_Notes_Rel1_0_0_0[]=
 {
     {   "Added a help button to URI input",
             "A new help button was added after the URI input on the main window.  "