+++ Tools/MakeReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,8 @@
 # Release Notes
 
+## Version 1.0.1.0 -- 16 Mar 2025
+ * Fixed bug with plugins with long descriptions
+   - The wrong buffer was used when loading a plugin with a description longer than 99 bytes
 ## Version 1.0.0.0 -- 01 Mar 2025
 ### First Release
  * Finished all features for first version